+++ .github/workflows/draft_pdf.yml
@@ -1,0 +1,27 @@
+name: Draft PDF
+on:
+  push:
+    paths:
+      - joos_paper/**
+      - .github/workflows/draft-pdf.yml
+jobs:
+  paper:
+    runs-on: ubuntu-latest
+    name: Rastereasy Paper Draft
+    steps:
+      - name: Checkout rastereasy paper
+        uses: actions/checkout@v4
+      - name: Build draft PDF
+        uses: openjournals/openjournals-draft-action@master
+        with:
+          journal: joss
+          # This should be the path to the paper within your repo.
+          paper-path: joos_paper/paper.md
+      - name: Upload
+        uses: actions/upload-artifact@v4
+        with:
+          name: paper
+          # This is the output path where Pandoc will write the compiled
+          # PDF. Note, this should be the same directory as the input
+          # paper.md
+          path: joos_paper/paper.pdf
