### 2025-06-21 b7682de
modifications joos paper
