+++ joos_paper/article.md
@@ -1,223 +1,0 @@
-
-----------
-
-title: 'Rastereasy: A Python package for an easy manipulation of remote sensing images'
-
-tags:
-
--   Python
--   Remote sensing
--   Geospatial analysis
--   Image processing
--   GIS
-
-
-authors:
-  -  name: Thomas Corpetti
-     orcid:  0000-0002-0257-138X
-     corresponding: true
-     affiliation: 1
-
-  -  name: Pierrick Matelot
-     affiliation: 2
-
-  -  name: Augustin de la Brosse
-     affiliation: 1
-
-  -  name: Candide Lissak
-     affiliation: 2
-     orcid:  0000-0003-3393-7767
-
-affiliations:
-  -  name: CNRS, UMR 6554 LETG, Univ. Rennes 2, Place du Recteur Henri Le Moal, 35043 Rennes Cedex, France
-     index: 1
-  - name: Université de Rennes, Inserm, Irset, UMR_S 1085
-     index: 2
-
-date: 03 June 2025
-bibliography: paper.bib
-
-----------
-
-# Summary
-
-The analysis and processing of remote sensing images have many important applications in various fields such as environmental monitoring, urban planning, or even agriculture. However, handling large georeferenced raster datasets can be challenging due to their complexity and size.
-
-**rastereasy** is a Python library for simple manipulation of georeferenced images (`*.tif`, `*.jp2`, `*.shp`, ...) [@ritter1997geotiff][@mamatov2024geospatial]. The goal is to simplify geospatial workflows by offering tools for reading and processing raster and vector files, resampling, cropping, reprojecting, stacking, etc of raster images, easy visualizations such as color composites and spectral plots, use (train / apply) some classical Machine Learning algorithms on images ...
-
-
-Compared to traditional RGB image manipulation, satellite images are highly specific due to their specific notions of spatial resolution, geographic extent, projection system, and they embed multiple spectral bands which prevents from an easy vizualisation. Dedicated software such as QGIS exists to handle these images, as well as specialized libraries for tasks like tiling, resampling, and reprojection. However, these tools require expertise in metadata management and geospatial systems, which can be a barrier for users unfamiliar with geographic data handling.
-
-**rastereasy** is designed to simplify these processes, providing an easy-to-use interface for standard operations on multispectral and georeferenced images. It is particularly aimed at users who are experienced in data processing but not necessarily in geospatial analysis, while also streamlining workflows for geographers by leveraging `rasterio` and other geospatial libraries. It is particularly useful, among other things, for preparing sample data for deep neural networks.
-
-The source code is available  at [https://easy-raster.github.io/](https://easy-raster.github.io/) and a documentation [here](https://github.com/pythonraster/rastereasy/).
-
-# Statement of need
-
-Many existing remote sensing libraries, such as `rasterio` and `gdal` [@garrard2016geoprocessing][@gillies2013rasterio], provide powerful functionalities but often require a deep understanding of geospatial data structures. `rastereasy` abstracts these complexities by offering a high-level interface for:
-
--   **Band manipulation**: Extract, reorder, and remove spectral bands easily.
-
--   **Tiling and stitching**: Split large raster images into smaller tiles and reconstruct them.
-
--   **Harmonization**: Align rasters with different spatial resolutions and extents.
-
--   **Visualization tools**: Quick and interactive display of georeferenced images and spectral signatures.
-
[…]
