+++ joos_paper/paper.md
@@ -13,10 +13,10 @@ authors:
   corresponding: true
   affiliation: 1
 - name: Pierrick Matelot
-  # orcid: XXXX-XXXX-XXXX-XXXX # Recommandé d'ajouter l'ORCID
+  orcid: 0000-0002-3932-0158
   affiliation: 2
 - name: Augustin de la Brosse
-  # orcid: XXXX-XXXX-XXXX-XXXX # Recommandé d'ajouter l'ORCID
+  orcid: 0000-0001-9793-9095
   affiliation: 1
 - name: Candide Lissak
   orcid: 0000-0003-3393-7767
@@ -31,6 +31,7 @@ affiliations:
 date: 03 June 2025
 bibliography: paper.bib
 ---
+
 
 # Summary
 
