### 2025-06-22 fdc747d
correction yaml reader
