### 2025-06-22 a94bd0d
correction yaml reader
