### 2025-06-23 584e45f
Update paper.md for compilation
