### 2025-06-23 a775fee
correction yaml reader
