### 2025-06-23 #8421 @thomas7619
The error is fixed and the paper compiles now, thank you !
