### 2025-06-25 #8421 @hugoledoux
I could but going on holidays 15 days from this Friday and no mobile reception (the ultimate holidays!)

If okay mid July then okay
