### 2025-07-28 #8421 @hugoledoux
@SimonMolinski would you be available for this review? I think it fits you
