### 2025-07-30 #8666 @editorialbot
[REVIEW]: Rastereasy: A Python package for an easy manipulation of remote sensing images
**Submitting author:** @thomas7619 (<a href="http://orcid.org/0000-0002-0257-138X">Thomas Corpetti</a>)
**Repository:** https://github.com/pythonraster/rastereasy/
**Branch with paper.md** (empty if default branch): main
**Version:** v0.3.2
