### 2025-08-13 #8666 @ArcticSnow
# JOSS review: Rastereasy

By @Arcticsnow - August 13, 2025

## General comments

Rastereasy is Python library written to facilitate the interaction with data stored in the form of raster (geospatial). It relies on well established libraries such as rasterio, shapely, and geopandas. Rastereasy offers an easy interface to interact with basic geospatial data (e.g. raster and shapefile). It provide access to GIS style commands in just few lines of code where most currently available Python libraries available have much verbose to accomplish the same actions. Therefore, typical GIS method are made easily accessible within a Python script. The library, to my opinion, fills a space where most existing Python library retain complexity, often not necessary to a user needing simple geospatial operations to be executed. With its current implementation, Rastereasy has a sound basis to be expanded in the future and incorporate more functionalities. To conclude, Rastereasy package is mainly a wrapper around the library Rasterio making use of a more human friendly grammar. It adds few functionalities like some bare bone ML algorithm inherited from Scikit-learn, raster clipping with shapefile, or shapefile rasterization. With this in mind, Rastereasy could find its place within the Python ecosystem of geospatial analysis.

Concerning the JOSS manuscript, it needs some attention and significant improvements. To my opinion, the statement of needs is poorly written. Throughout the manuscript there is an intense usage of "..." and uncomplete sentences. The manuscript includes much code snippets in the Example section that bring little to the manuscript and is already well described in the documentation. The section on Performance falls short to demonstrate or indicate limitations of the toolbox in its current implementation (e.g. memory management). Finally as it is more of a wrapper style library, it would be important to state and reference the core libraries on which it depends on. I recommend a deep rewriting and editing of the manuscript for better clarity and language.

## Specific comments

**Documentation**: Very clear

**Coding experience**:

- Plotting in a simple Ipython console is not straight forward as to obtain a simple Pyplot interactive window.
- Source code is well commented. Here and there comments are in French.

**Community Guideline**: please add guidelines as to how someone ma come and contribute to the project.

**Other comments**:

- line 8-9: I find this sentence vague adding little information. "... many important ... various field ..."
- line 12: should read: "simple manipulation of georeferenced datasets" (.shp are no images)
- line 13: citations are poorly formated.
- line 16: no ending to sentence. This feels straight out of a LLM
- line 44: remove ... or specify
- line 56: add  capital and dot to sentence to format it as a sentence properly.
- line 60: I think you meant "if" instead of "in"
- Example for Harmonization could show a figure before after to illustrate. Otherwise, I see little interest in providing the example in th manuscript.
- In Performance and Scalability, I wonder how is rastereasy dealing in term of memory. Can it make use of lazy loading or does it need to load the whole image image into memory? As some satelite images can be quite large, this may be important for a user to consider if handling large datasets.
