### 2025-08-14 acccaba
revised version
