+++ joos_paper/paper.bib
@@ -2,16 +2,20 @@
   title={Rasterio: Geospatial raster I/O for Python programmers},
   author={Gillies, Sean and Ward, B and Petersen, AS and others},
   journal={URL https://github. com/mapbox/rasterio},
+  doi={},
   year={2013}
 }
+
 @incollection{mamatov2024geospatial,
   title={Geospatial Data Storage and Management},
   author={Mamatov, Islom and Galety, Mohammad Gouse and Alimov, Ruslan and Sriharsha, AV and Rofoo, Fanar Fareed Hanna and Sunitha, Gurram},
   booktitle={Ethics, Machine Learning, and Python in Geospatial Analysis},
   pages={150--167},
   year={2024},
-  publisher={IGI Global}
+  doi={10.4018/979-8-3693-6381-2.ch007},
+  publisher={IGI Global Scientific Publishing}
 }
+
 
 
 @article{ritter1997geotiff,
@@ -22,6 +26,7 @@
   number={7},
   pages={1637--1647},
   year={1997},
+  doi={10.1080/014311697218340},
   publisher={Taylor \& Francis}
 }
 @article{xue2017significant,
@@ -32,6 +37,7 @@
   number={1},
   pages={1353691},
   year={2017},
+  doi={10.1155/2017/1353691},
   publisher={Wiley Online Library}
 }
 @inproceedings{courty2016optimal,
@@ -40,6 +46,7 @@
   booktitle={2016 IEEE international geoscience and remote sensing symposium (IGARSS)},
   pages={3571--3574},
   year={2016},
+  doi={10.1109/IGARSS.2016.7729925},
   organization={IEEE}
 }
 
@@ -50,6 +57,7 @@
   volume={622},
   pages={178--210},
   year={2023},
+  doi={10.1016/j.ins.2022.11.139},
   publisher={Elsevier}
 }
 
@@ -60,6 +68,44 @@
   volume={1},
   pages={330--331},
   year={1992}
+}
+
+@article{harris2020array,
+  title={Array programming with NumPy},
+  author={Harris, Charles R and Millman, K Jarrod and Van Der Walt, St{\'e}fan J and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J and others},
+  journal={nature},
+  volume={585},
+  number={7825},
+  pages={357--362},
+  year={2020},
+  doi={10.1038/s41586-020-2649-2},
+  publisher={Nature Publishing Group UK London}
+}
+
+
+@article{gillies2013shapely,
+  title={The shapely user manual},
+  author={Gillies, Sean and others},
+  journal={URL https://pypi. org/project/Shapely},
+  year={2013}
+}
+
+@article{jordahl2021geopandas,
+  title={geopandas/geopandas: v0. 7.0},
+  author={Jordahl, Kelsey and Van den Bossche, Joris and Wasserman, Jacob and McBride, James and Fleischmann, Martin and Gerard, Jeffrey and Tratner, Jeff and Perry, Matthew and Farmer, Carson and Hjelle, Geir Arne and others},
+  journal={Zenodo},
+  doi={10.5281/zenodo.3669853},
+  year={2021}
+}
+
+
+@incollection{kramer2016scikit,
+  title={Scikit-learn},
+  author={Kramer, Oliver},
+  booktitle={Machine learning for evolution strategies},
+  pages={45--53},
+  year={2016},
+  publisher={Springer}
 }
 
 
