+++ joos_paper/paper.md
@@ -29,100 +29,77 @@ bibliography: paper.bib
 
 # Summary
 
-The analysis and processing of remote sensing images have many important applications in various fields such as environmental monitoring, urban planning, or even agriculture. However, handling large georeferenced raster datasets can be challenging due to their complexity and size.
+Working with remote sensing data often involves managing large, multi-band georeferenced rasters with varying spatial resolutions, extents, and coordinate reference systems [@mamatov2024geospatial]. Established libraries such as `rasterio` and `GDAL` [@garrard2016geoprocessing; @gillies2013rasterio] provide extensive capabilities for these tasks, but they can be verbose and require a solid understanding of geospatial concepts such as projections, geotransforms, and metadata handling. For users whose primary expertise lies outside GIS—such as data scientists, ecologists, agronomists, or climate researchers—this steep learning curve can hinder the rapid development of operational workflows.
 
-**rastereasy** is a Python library for simple manipulation of georeferenced images (`*.tif`, `*.jp2`, `*.shp`, ...) [@ritter1997geotiff][@mamatov2024geospatial]. The goal is to simplify geospatial workflows by offering tools for reading and processing raster and vector files, resampling, cropping, reprojecting, stacking, etc of raster images, easy visualizations such as color composites and spectral plots, use (train / apply) some classical Machine Learning algorithms on images ...
+**rastereasy** is a Python library designed to bridge this gap by providing a high-level, human-readable interface for common geospatial raster and vector operations (e.g., *.tif, *.jp2, *.shp) [@ritter1997geotiff; @mamatov2024geospatial]. Built on well-established libraries including `rasterio`, `numpy`, `shapely`, `geopandas`, and `scikit-learn` [@gillies2013rasterio; @harris2020array; @gillies2013shapely; @jordahl2021geopandas; @kramer2016scikit], it enables users to perform typical GIS tasks—such as resampling, cropping, reprojection, stacking, clipping rasters with shapefiles, or rasterizing vector layers—in just a few lines of code. Some basic Machine Learning functionalities (clustering, fusion) are also implemented.
+
+By abstracting away much of the underlying technical complexity, **rastereasy** makes geospatial processing directly accessible within Python scripts. It is particularly suited for analysts and machine learning practitioners who need to integrate geospatial data handling into their workflows without deep GIS expertise, while also helping experienced geographers prototype more quickly. Beyond core raster operations, it includes utilities for harmonizing multi-source imagery, performing clustering and domain adaptation, and preparing datasets for downstream analysis.
 
 
-Compared to traditional RGB image manipulation, satellite images are highly specific due to their specific notions of spatial resolution, geographic extent, projection system, and they embed multiple spectral bands which prevents from an easy vizualisation. Dedicated software such as QGIS exists to handle these images, as well as specialized libraries for tasks like tiling, resampling, and reprojection. However, these tools require expertise in metadata management and geospatial systems, which can be a barrier for users unfamiliar with geographic data handling.
+With its current implementation, **rastereasy** provides a solid foundation for further development and integration into the Python geospatial ecosystem. The source code is available at [https://github.com/pythonraster/rastereasy](https://github.com/pythonraster/rastereasy) and a documentation [https://rastereasy.github.io/](https://rastereasy.github.io/).
 
[…]
