### 2025-08-14 #8666 @thomas7619
Thank you very much @ArcticSnow  for your detailed feedback and for taking the time to review our manuscript. We acknowledge the points you raised and have made the following changes:
- Statement of need

Completely rewritten for clarity and structure, avoiding incomplete sentences and the use of “...”. We now  explain the motivation, the target audience, and the position of Rastereasy within the Python geospatial ecosystem.

- Code snippets in the Example section

We have reduced the number of code snippets presented in the manuscript. We now include only concise, representative code blocks that illustrate the main functionalities of the package. Additionally, we have added a visual example for the harmonization functionality to more clearly demonstrate the practical value of the package. For readers interested in more comprehensive examples and tutorials, we refer them to the full online documentation available on our GitHub repository.

- Performance section
We have described current limitations, particularly regarding memory usage. We clarified that, by default, the library loads full rasters into memory; however, windowed reading via rasterio could be supported for handling large datasets efficiently. We also highlighted that features such as lazy loading and parallel processing are planned for future releases to further improve performance and scalability.

- References to core libraries
Regarding references to core libraries, we have explicitly stated in both the Statement of Need and the Summary that rastereasy builds upon rasterio, shapely, geopandas, and scikit-learn. Corresponding citations for these libraries have also been added to the bibliography to properly acknowledge their contributions.

- Language and clarity
In response to the reviewer’s comment on language and clarity, we have edited the entire manuscript to ensure consistency and improve readability. Incomplete sentences (issued from cut errors and not LLM !) have been removed, and formatting issues have been corrected throughout the text.

We hope these revisions address the main concerns raised and improve both the clarity and the scientific value of the manuscript. Thank you again for all these details
