### 2025-08-25 4470970
add automated tests
