+++ tests/test_core.py
@@ -1,0 +1,458 @@
+
+import rastereasy
+import numpy as np
+
+
+def test_all():
+    name_im='./tests/data/sentinel.tif'
+    Image=rastereasy.Geoimage(name_im)
+    names = {"NIR":8,"G":3,"CO" : 1,"SWIR2":11,"B": 2,"R":4,"RE1":5,"RE2":6,"RE3":7,"WA":9,"SWIR1":10,"SWIR3":12}
+    Image_names=rastereasy.Geoimage(name_im,names=names)
+    Image_names.crop(100,150,100,150,inplace=True)
+    im1=Image.crop(100,150,100,150)
+    im2=Image.crop(250,300,250,300)
+    diff=im2-im1
+    diff2=diff.abs()
+    diff.abs(inplace=True)
+    assert (diff-diff2).abs().sum() == 0
+
+    Image.info()
+    Image_names.get_meta()['dtype']
+    row=20
+    col=10
+    pixel_row = 15
+    pixel_col = 22
+    assert np.all((Image[row,:][:,pixel_col]-Image[row,pixel_col]) == 0)
+    assert np.all((Image[:,col][:,pixel_row]-Image[pixel_row,col]) == 0)
+
+
+    Image_std,scaler_std=Image.standardize(dest_name='./tests/res_test/test_image_S2_std.tif')
+    Image_minmax,scaler_minmax=Image.standardize(dest_name='./tests/res_test/test_image_S2_minmax.tif',type='minmax')
+
+
+    Image=rastereasy.Geoimage(name_im)
+    Image_apply_std=Image.standardize(scaler_std)
+    Image_apply_minmax=Image.standardize(scaler_minmax)
+    assert (np.sum(np.abs(Image_apply_std.image-Image_std.image))) ==0
+    assert(np.sum(np.abs(Image_apply_minmax.image-Image_minmax.image)))==0
+
+
+    Image_recovered_std=Image_std.inverse_standardize(scaler_std)
+    Image_recovered_minmax=Image_minmax.inverse_standardize(scaler_minmax)
+
+    Image=rastereasy.Geoimage(name_im)
+    Image.standardize(inplace=True)
+    assert((Image_std-Image).abs().sum())==0
+
+    Image=rastereasy.Geoimage(name_im)
+
+    Image.standardize(type='minmax',inplace=True)
+    assert((Image_minmax-Image).abs().sum())==0
+
+    name_im='./tests/data/sentinel.tif'
+    Image=rastereasy.Geoimage(name_im)
+    Image.standardize(scaler_std,
+                      dest_name='./tests/res_test/test_image_S2_std2.tif',
+                      inplace=True)
+    assert((Image_std-Image).abs().sum())==0
+
+    Image=rastereasy.Geoimage(name_im)
+    Image.standardize(scaler_minmax,
+                      dest_name='./tests/res_test/test_image_S2_minmax2.tif',
+                      inplace=True)
+    assert((Image_minmax-Image).abs().sum()) ==0
+
+    name_im='./tests/res_test/test_image_S2_minmax2.tif'
+    I1=rastereasy.Geoimage(name_im)
+    name_im='./tests/res_test/test_image_S2_minmax.tif'
+    I2=rastereasy.Geoimage(name_im)
+    print((I1-I2).abs().sum())
+    assert((I1-I2).abs().sum()) ==0
+
+    name_im='./tests/res_test/test_image_S2_std.tif'
+    I1=rastereasy.Geoimage(name_im)
+    name_im='./tests/res_test/test_image_S2_std2.tif'
+    I2=rastereasy.Geoimage(name_im)
[…]
