### 2025-08-25 9adb577
add automated tests
