+++ .github/workflows/tests.yml
@@ -1,0 +1,25 @@
+name: Run tests
+
+on: [push, pull_request]
+  paths:
+    - 'tests/**'
+    - '.github/workflows/tests.yml'
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.10'
+
+      - name: Install dependencies
+        run: |
+          pip install -e .
+          pip install pytest rastereasy 
+
+      - name: Run tests
+        run: pytest -v
