### 2025-08-25 d3fab0d
add automated tests
