+++ .github/workflows/tests.yml
@@ -1,10 +1,14 @@
 name: Run tests
 
-on: [push, pull_request]
-  paths:
-    - 'tests/**'
-    - '.github/workflows/tests.yml'
-
+on:
+  push:
+    paths:
+      - 'tests/**'
+      - '.github/workflows/tests.yml'
+  pull_request:
+    paths:
+      - 'tests/**'
+      - '.github/workflows/tests.yml'
 jobs:
   test:
     runs-on: ubuntu-latest
@@ -19,7 +23,7 @@ jobs:
       - name: Install dependencies
         run: |
           pip install -e .
-          pip install pytest rastereasy 
+          pip install pytest rastereasy
 
       - name: Run tests
         run: pytest -v
