### 2025-08-25 #8666 @thomas7619
Dear @ArcticSnow. Thanks a lot for this feedback, and for your positive points.
As you suggested, I have now added automated unit tests so that each push will test functions of the librery.
Feel free to contribute to the library
