### 2025-08-29 #8666 @hugoledoux
> I am not fully competent in the matter, and so I'll leave to [@hugoledoux](https://github.com/hugoledoux) judging and informing further on this topic.

@ArcticSnow JOSS simply requires that there are at least unit tests and that automatically be used to test if changes break the functionalities. If you can run those and they look logical to you, then you can click that checkbox.
