### 2025-09-06 7f8341a
add precision on colab
