+++ src/rastereasy/rastereasy.py
@@ -48,7 +48,12 @@ import geopandas as gpd
 
 # Local imports with relative path
 from .utils import *
-
+from scipy.ndimage import (
+    gaussian_filter,
+    median_filter,
+    laplace,
+    sobel
+)
 # Configure warnings
 warnings.filterwarnings("ignore", category=DeprecationWarning)
 # Configure matplotlib backend based on environment
@@ -1738,11 +1743,13 @@ def extract_common_areas(im1, im2, resolution='min', projection=None):
     lat2_d, lon2_d = im2.pixel2latlon(0, 0)
     lat2_e, lon2_e = im2.pixel2latlon(im2.shape[0], im2.shape[1])
 
+
     # Calculate boundaries of overlapping area
     bound_lat_d = min(lat1_d, lat2_d)
     bound_lat_e = max(lat1_e, lat2_e)
     bound_lon_d = max(lon1_d, lon2_d)
     bound_lon_e = min(lon1_e, lon2_e)
+
 
     # Crop images to common bounds
     im1common = im1.crop(bound_lon_d, bound_lon_e, bound_lat_d, bound_lat_e, pixel=False)
@@ -5182,7 +5189,7 @@ class Geoimage:
         meta['dtype'] = str(image.dtype)
 
         # Create the new Geoimage
-        geoim = Geoimage(data=image, meta=meta, names=names, georef=self.__georef)
+        geoim = Geoimage(data=image, meta=meta, names=names, georef=self.__georef, history=self.__history)
 
         # Save if requested
         if dest_name is not None:
@@ -8952,3 +8959,223 @@ class Geoimage:
 
         except Exception as e:
             raise RuntimeError(f"Creating image with additional band failed: {str(e)}") from e
+            
+    def __generic_filter(self, kernel, inplace = False, dest_name=None):
+        """
+        Apply a generic 2D convolution filter to the spectral bands of the image.
+    
+        This method applies a user-defined kernel (2D array) to each spectral band 
+        of the image using convolution. It can be used, for example, to perform 
+        smoothing (mean filter), edge detection, or other custom spatial filtering 
+        operations.
+    
+        Parameters
+        ----------
+        kernel : numpy.ndarray
+            A 2D convolution kernel. Must be a square or rectangular matrix 
+            (e.g., a Gaussian blur kernel, Sobel operator, etc.).
+    
+        inplace : bool, default False
+            If False, returns a new Geoimage instance with the filtered data.
+            If True, modifies the current image in place.
+    
+        dest_name : str, optional
+            Path to save the filtered image. If None, the image is not saved.
+            Default is None.
+    
+        Returns
+        -------
+        Geoimage or None
+            A new Geoimage containing the filtered image if `inplace=False`.
+            Returns None if `inplace=True`.
+    
+        Raises
+        ------
+        ValueError
+            If the kernel dimensions are invalid or if the image data type 
+            is not supported.
+    
+        Examples
+        --------
+        >>> # Create an average filter of size 5
+        >>> blur_kernel = np.ones((5, 5)) / 25
+        >>> # Apply the filter and return a new Geoimage
+        >>> Image_filtered = Image.generic_filter(blur_kernel)
+        >>> # Apply the filter in place and save the result
+        >>> Image.generic_filter(blur_kernel, inplace=True, dest_name="im_filtered.tif")
+    
+        Notes
+        -----
+        - The kernel is applied independently to each spectral band.
+        - This function uses convolution; kernel normalization (e.g., sum to 1) 
+          is the responsibility of the user.
+        - For large images, filtering may require significant memory.
+        """
+        
+        blurred_image = apply_filter(self.numpy_channel_last().astype(np.float64), kernel)
+        if inplace:
+            self.upload_image(blurred_image,channel_first=False, inplace = True)
+            if dest_name is not None:
[…]
