+++ src/rastereasy.egg-info/SOURCES.txt
@@ -2,14 +2,19 @@ LICENCE
 README.md
 environment_rastereasy.yml
 pyproject.toml
+.github/workflows/draft_pdf.yml
+.github/workflows/tests.yml
+dist/rastereasy-0.1.1-py3-none-any.whl
+dist/rastereasy-0.1.1.tar.gz
 illus/spectra.png
-joos_paper/.DS_Store
+joos_paper/harmonization.jpg
+joos_paper/illustrations.jpg
 joos_paper/illustrations.pdf
+joos_paper/illustrations2.jpg
 joos_paper/paper.bib
 joos_paper/paper.md
 src/.DS_Store
 src/rastereasy/.DS_Store
-src/rastereasy/Untitled.ipynb
 src/rastereasy/__init__.py
 src/rastereasy/rastereasy.py
 src/rastereasy/utils.py
@@ -23,3 +28,16 @@ src/rastereasy/.ipynb_checkpoints/rastereasy-checkpoint.py
 src/rastereasy/.ipynb_checkpoints/utils-checkpoint.py
 src/rastereasy/__pycache__/rastereasy.cpython-311.pyc
 src/rastereasy/__pycache__/utils.cpython-311.pyc
+tests/test_core.py
+tests/data/RGB_common_1.tif
+tests/data/RGB_common_2.tif
+tests/data/sentinel.tif
+tests/shp/Roi_G5.cpg
+tests/shp/Roi_G5.dbf
+tests/shp/Roi_G5.prj
+tests/shp/Roi_G5.shp
+tests/shp/Roi_G5.shx
+tests/tostack/G5_B2.tif
+tests/tostack/G5_B3.tif
+tests/tostack/G5_B4.tif
+tests/tostack/G5_B8.tif
