+++ src/rastereasy.egg-info/SOURCES.txt
@@ -4,8 +4,6 @@ environment_rastereasy.yml
 pyproject.toml
 .github/workflows/draft_pdf.yml
 .github/workflows/tests.yml
-dist/rastereasy-0.1.1-py3-none-any.whl
-dist/rastereasy-0.1.1.tar.gz
 illus/spectra.png
 joos_paper/harmonization.jpg
 joos_paper/illustrations.jpg
