+++ src/rastereasy/rastereasy.py
@@ -666,6 +666,65 @@ def table2image(table, size, channel_first=True):
         image = table.reshape(size[0], size[1], bands)
         return image
 
+
+def extract_numpy_tables(data, outputs, label=None):
+    """
+    Extract paired NumPy tables (features and outputs) for machine learning.
+
+    This function converts raster data stored in `Geoimage` objects into two
+    NumPy arrays: one containing the input features (X), and one containing
+    the corresponding outputs/labels (y). Optionally, it can filter the
+    dataset to only include samples corresponding to one or multiple label
+    values.
+
+    Parameters
+    ----------
+    data : Geoimage
+        A `rastereasy.Geoimage` containing the input features
+        (e.g., multispectral bands).
+    outputs : Geoimage
+        A `rastereasy.Geoimage` containing the outputs/labels
+        (e.g., classes or quantitative values).
+    label : int or list of int, optional (default=None)
+        If provided, only the samples corresponding to this label value
+        (or list of values) in `outputs` will be extracted.
+
+    Returns
+    -------
+    X : numpy.ndarray
+        Input feature table of shape (N, f), where
+        N is the number of extracted samples and
+        f is the number of features (bands).
+    y : numpy.ndarray
+        Output array of shape (N, ), containing the labels/outputs
+        associated with each sample.
+
+    Examples
+    --------
+    >>> # Extract all data/labels from two Geoimages
+    >>> X, y = extract_numpy_tables(data, labels)
+
+    >>> # Extract only the samples where label == 1
+    >>> X, y = extract_numpy_tables(data, outputs, label=1)
+
+    >>> # Extract only the samples with labels in [1, 3, 5]
+    >>> X, y = extract_numpy_tables(data, outputs, label=[1, 3, 5])
+    """
+    data_np = data.numpy_table()
+    classes_np = outputs.numpy_table()
+
+    if label is None:
+        X = data_np
+        y = classes_np
+    else:
+        # Assure qu'on gère int et liste
+        labels = np.atleast_1d(label)
+        mask = np.isin(classes_np.flatten(), labels)
+        X = data_np[mask]
+        y = classes_np.flatten()[mask]
+    return X, y
+
+
 def shp2geoim2(shapefile_path, attribute='code', resolution=10, nodata=0):
     """
     Convertit un shapefile en données raster et métadonnées géospatiales.
@@ -3179,7 +3238,7 @@ class Geoimage:
             now_str = now.strftime("%Y-%m-%d %H:%M:%S")
             self.__listhistory.append(f'[{now_str}] - Changed band names')
 
-        return self
+        # return self
 
 
     def activate_history(self):
@@ -3607,6 +3666,7 @@ class Geoimage:
             - 'band': Sum across spectral bands for each pixel
             - 'row': Sum across rows (lines) for each band and column
             - 'col': Sum across columns for each band and row
+            - 'pixel': Sum across pixels for each bands
             - None: Sum of all values in the image
             Default is None.
 
@@ -3658,6 +3718,7 @@ class Geoimage:
             - 'band': Minimum across spectral bands for each pixel
             - 'row': Minimum across rows (lines) for each band and column
             - 'col': Minimum across columns for each band and row
+            - 'pixel': Minimum across pixels for each bands
             - None: Global minimum of the entire image
             Default is None.
 
@@ -3709,6 +3770,7 @@ class Geoimage:
             - 'band': Maximum across spectral bands for each pixel
             - 'row': Maximum across rows (lines) for each band and column
             - 'col': Maximum across columns for each band and row
+            - 'pixel': Maximum across pixels for each bands
             - None: Global maximum of the entire image
             Default is None.
 
@@ -3760,6 +3822,7 @@ class Geoimage:
             - 'band': Mean across spectral bands for each pixel
[…]
