+++ src/rastereasy/rastereasy.py
@@ -25,6 +25,7 @@ from itertools import product
 import warnings
 import copy
 import datetime
+import json
 
 # Scientific and numerical libraries
 import numpy as np
@@ -69,6 +70,11 @@ DEF_FIG_SIZE = (5, 5)  # Default figure size for visualizations
 RANDOM_STATE = None    # Random state for reproducible results
 
 
+        
+
+
+
+
 def read_geoim(source_name, read_image=True, channel_first=True):
     """
     Read a geotiff image file.
@@ -92,26 +98,35 @@ def read_geoim(source_name, read_image=True, channel_first=True):
         Image data if read_image is True, None otherwise
     dict
         Metadata dictionary from rasterio
+    names
+        Metadata dictionary with names of the bands (if given)
 
     Examples
     --------
     >>> # Read image data and metadata
-    >>> image, meta = read_geoim("path/to/image.tif")
+    >>> image, meta, names = read_geoim("path/to/image.tif")
     >>>
     >>> # Read only metadata
-    >>> meta = read_geoim("path/to/image.tif", read_image=False)
+    >>> meta, names = read_geoim("path/to/image.tif", read_image=False)
     """
     src = rio.open(source_name)
+    extra_tags = None
+    tags = src.tags()
+    if "EXTRA_TAGS" in tags:
+        try:
+            extra_tags = json.loads(tags["EXTRA_TAGS"])
+        except json.JSONDecodeError:
+            extra_tags = tags["EXTRA_TAGS"]
     if read_image:
         if channel_first:
-            return src.read(), src.meta
+            return src.read(), src.meta, extra_tags
         else:
-            return np.rollaxis(src.read(), 0, 3), src.meta
+            return np.rollaxis(src.read(), 0, 3), src.meta, extra_tags
     else:
-        return src.meta
+        return src.meta, extra_tags
 
 
-def write_geoim(im, meta_input, dest_name, channel_first=True):
+def write_geoim(im, meta_input, dest_name, channel_first=True, names=None):
     """
     Write a geotiff image to disk.
 
@@ -123,6 +138,8 @@ def write_geoim(im, meta_input, dest_name, channel_first=True):
         Metadata dictionary with rasterio metadata fields
     dest_name : str
         Output path for the geotiff file
+    names : dict
+        Names (in a dict) associated with spectral bands
     channel_first : bool, optional
         If True, assumes image has shape (bands, rows, cols).
         If False, assumes image has shape (rows, cols, bands).
@@ -160,10 +177,13 @@ def write_geoim(im, meta_input, dest_name, channel_first=True):
     if channel_first:
         with rio.open(dest_name, 'w', **meta) as dst:
             dst.write(im)
+            if names is not None:
+                dst.update_tags(EXTRA_TAGS=json.dumps(names))
     else:
         with rio.open(dest_name, 'w', **meta) as dst:
             dst.write(np.rollaxis(im, 2, 0))
-
+            if names is not None:
+                dst.update_tags(EXTRA_TAGS=json.dumps(names))
 
 def diff_im(im1, im2):
     """
@@ -311,7 +331,7 @@ def im2tiles(source_name, dest_name, nb_row, nb_col, overlap=0, type_name='seque
         name_tile=name_tile
     )
 
-def crop_rio(data, deb_row, end_row, deb_col, end_col, dest_name=None, meta=None, channel_first=True):
+def crop_rio(data, deb_row, end_row, deb_col, end_col, dest_name=None, meta=None, channel_first=True, names=None):
     """
     Crop a georeferenced image.
 
@@ -333,6 +353,9 @@ def crop_rio(data, deb_row, end_row, deb_col, end_col, dest_name=None, meta=None
         Default is None.
     meta : dict, optional
[…]
