### 2025-09-19 #8666 @thomas7619
Dear @SimonMolinsky and @hugoledoux,
I’m not sure whether I have anything to do at this stage regarding @SimonMolinsky’s checklist.
FYI, I also updated the library with additional functionalities (reading only metadata, reading parts of images, adding filtering functions, saving band names with the image, etc.).
Thanks a lot,
Thomas
