### 2025-09-21 #8666 @SimonMolinsky
Hi @thomas7619

I appreciate your work! There is enormous potential in simplifying raster manipulations and using ML models on satellite imagery. Designing ETL and ML pipelines for raster data takes a lot of time and `rastereasy` is the solution for this problem!

The package has clear documentation. Installation is straightforward (MacOS), and API structure is logical and well described, but there are some minor issues (see comments below). Functionalities work as expected. I see a great value in functions describing band statistics and simplifying calculations as well ML models development. You have provided data used in your examples, thus your output is reproducible. Examples are thorough, and show package capabilities. Package code follows good practices, but I would move functionalities to submodules instead of having every class and function in a single file (but it’s up to you).

## Comments regarding PAPER

**Performance and Scalability**: you write about the future development, but I'd rather skip this part of the paper and focus on the current performance only. You might add one short sentence that you plan to address current limitations in the future and that's all.

**State of the field**: I think that some similar packages are missing. rioxarray, Raster Forge, PODPAC, EarthPy. Could you compare those packages to your software?

## Comments regarding PACKAGE

- Some methods are described using verbs in present continuous / or gerunds (`resampling()`) and others as verbs in present simple (`reproject()`). Standard approach is the second one, please, consider changing methods and functions to one standardized naming schema.
- You don’t have clear community guidelines (or I didn't notice where they are). Nice tutorial how to prepare community guidelines is here: https://www.pyopensci.org/python-package-guide/documentation/repository-files/contributing-file.html (This is related to the point: `Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support`) You can update your README with this information if you want to instead of creating new files.
- Code of conduct in your docs point to the French language version, you might consider using English (user can change language of this page, so it's up to you).

## OPTIONAL comments, not required for the review acceptance

- bump the minimal Python version to 3.9 in pyproject.toml because 3.7 and 3.8 are not supported (see the graph here: https://devguide.python.org/versions/)
- Clean repo from [`__pycache__, .ipynb_checkpoints, .DS_Store`] files
- Naming conventions: `apply_ML_model`  is self-explanatory, but it doesn't follow industry standards. Have you considered different function names/aliases for this endpoint like: `predict`, `inference`, or `model` that are closer to the standard naming conventions?
