+++ src/rastereasy/rastereasy.py
@@ -70,7 +70,7 @@ DEF_FIG_SIZE = (5, 5)  # Default figure size for visualizations
 RANDOM_STATE = None    # Random state for reproducible results
 
 
-        
+
 
 
 
@@ -355,7 +355,7 @@ def crop_rio(data, deb_row, end_row, deb_col, end_col, dest_name=None, meta=None
         Metadata dictionary (required if data is a numpy array).
         Default is None.
     names : dict, optional
-        Band name dictionary 
+        Band name dictionary
         Default is None.
     channel_first : bool, optional
         If True, assumes/returns image with shape (bands, rows, cols).
@@ -538,7 +538,7 @@ def resampling(data, final_resolution, dest_name=None, method='cubic_spline', ch
         Metadata to use if `data` is a numpy array.
         Default is None.
     names : dict, optional
-        Band name dictionary 
+        Band name dictionary
         Default is None.
 
     Returns
@@ -1693,7 +1693,7 @@ def files2stack(imagefile_path, resolution=None, names="origin", dest_name=None,
             if ims.nb_bands!=1:
                 message="im %s has %d bands"%(imagefile_path[i + 1],ims.nb_bands)
                 warnings.warn(message,category=UserWarning)
-                
+
             im.stack(ims, reformat_names=True, inplace = True)
 
         # Handle band naming
@@ -2084,7 +2084,7 @@ def colorcomp(image, bands, name_save='', names=None, percentile=2, channel_firs
     im[:, :, 1] = normalize(image[:, :, names[bands[1]] - 1], percentile)
     im[:, :, 2] = normalize(image[:, :, names[bands[2]] - 1], percentile)
 
-    
+
 
 # {'R': 1, 'TGT': 2, 'EZS': 3}
 
@@ -2137,67 +2137,67 @@ class Geoimage:
     This class provides a comprehensive toolkit for working with geospatial raster data,
     supporting operations such as image creation, visualization, band manipulation,
     reprojection, resampling, cropping, and more.
-    
+
     """
-    
+
 
 
     def __init__(self, source_name=None, meta_only=False,names=None, history=False,
                  data=None, meta=None, georef=None, target_crs="EPSG:4326",area=None,extent='pixel'):
         """
         Initialize a Geoimage object from a file or data array with metadata.
-        
+
         Parameters
         ----------
         source_name : str, optional
-            Path to a geoimage (.tif, .jp2) image file to load. 
-            If provided, the image data and metadata 
+            Path to a geoimage (.tif, .jp2) image file to load.
+            If provided, the image data and metadata
             will be read from this file.
         meta_only : bool, optional
             If True, do not read the image but just
             the meta information (useful for image.info()).
         names : dict, optional
-            Dictionary mapping band names to 
+            Dictionary mapping band names to
             band indices (e.g., {'NIR': 1, 'R': 2, 'G': 3}).
-            If not provided, bands will be 
+            If not provided, bands will be
             named numerically ('1', '2', '3', ...).
         area : tuple, optional
             To read  only a window of the image
-                If based on pixel coordinates, you must indicate 
-                - the row/col coordinades of 
+                If based on pixel coordinates, you must indicate
+                - the row/col coordinades of
                         the north-west corner (deb_row,deb_col)
-                - the row/col coordinades of 
+                - the row/col coordinades of
                         the south-east corner (end_row,end_col)
                 in a tuple  `area = ((deb_row,end_row),(deb_col,end_col))`
-                
-                If based on latitude/longitude coordinates, you must indicate 
+
+                If based on latitude/longitude coordinates, you must indicate
                 - the lat/lon coordinades of the north-west corner (lat1,lon1)
[…]
