+++ joos_paper/paper.bib
@@ -116,3 +116,42 @@
   year={2016},
   publisher={Simon and Schuster}
 }
+
+@article{oliveira2024raster,
+  title={Raster Forge: Interactive raster manipulation library and GUI for Python},
+  author={Oliveira, Afonso and Fachada, Nuno and Matos-Carvalho, Jo{\~a}o P},
+  journal={Software Impacts},
+  volume={20},
+  pages={100657},
+  year={2024},
+  publisher={Elsevier}
+}
+
+@article{ueckermann2020podpac,
+  title={PODPAC: open-source Python software for enabling harmonized, plug-and-play processing of disparate earth observation data sets and seamless transition onto the serverless cloud by earth scientists},
+  author={Ueckermann, Mattheus P and Bieszczad, Jerry and Entekhabi, Dara and Shapiro, Marc L and Callendar, David R and Sullivan, David and Milloy, Jeffrey},
+  journal={Earth Science Informatics},
+  volume={13},
+  number={4},
+  pages={1507--1521},
+  year={2020},
+  publisher={Springer}
+}
+
+@article{Wasser2019EarthPy,
+	journal = {Journal of Open Source Software},
+	doi = {10.21105/joss.01886},
+	issn = {2475-9066},
+	number = {43},
+	publisher = {The Open Journal},
+	title = {EarthPy: A Python package that makes it easier to explore and plot raster and vector data using open source Python tools.},
+	url = {https://doi.org/10.21105/joss.01886},
+	volume = {4},
+	author = {Wasser, Leah and Joseph, Maxwell and McGlinchy, Joe and Palomino, Jenny and Korinek, Nathan and Holdgraf, Chris and Head, Tim},
+	pages = {1886},
+	date = {2019-11-13},
+	year = {2019},
+	month = {11},
+	day = {13},
+}
+
