+++ CONTRIBUTING.md
@@ -1,0 +1,31 @@
+# Contributing to rastereasy
+
+We're happy that you're interested in contributing to `rastereasy`! All contributions, whether it's fixing bugs, improving documentation, or proposing new features, are welcome.
+
+To ensure a smooth and effective process for everyone, please follow these guidelines.
+
+### **Reporting Bugs or Suggesting Enhancements**
+
+The easiest way to report a bug or suggest a new feature is to open an **issue** on our [GitHub repository](https://github.com/pythonraster/rastereasy/issues).
+
+* **For a bug report**: Please describe the issue as precisely as possible. Include the `rastereasy` version you are using, a short, reproducible code snippet that demonstrates the problem, the result you got, and what you expected to get.
+
+* **For a feature suggestion**: Please clearly explain the feature you would like to see added and why it would be useful for the project.
+
+### **Proposing Changes (Pull Requests)**
+
+If you want to contribute directly to the code or documentation, we use GitHub's **Pull Request** (PR) system.
+
+1.  **Fork the project**: Create a copy ("fork") of the [`rastereasy` repository](https://github.com/pythonraster/rastereasy/) to your own GitHub account.
+
+2.  **Create a branch**: On your fork, create a new branch dedicated to your changes (e.g., `fix-bug-xyz` or `feature-new-function`).
+
+3.  **Make your changes**: Implement your changes in this branch. Feel free to add tests if relevant.
+
+4.  **Submit a Pull Request**: Once your changes are complete, open a Pull Request from your branch to the `main` branch of the `rastereasy` project. Clearly describe the changes you've made in the PR description.
+
+We will review your contribution as soon as possible!
+
+### **Seeking Support**
+
+If you have questions about using `rastereasy` or need help, the best way to reach us is by opening an **issue** on GitHub and describing your question. This allows the entire community to benefit from the answers.
