+++ joos_paper/paper.bib
@@ -124,6 +124,7 @@
   volume={20},
   pages={100657},
   year={2024},
+  doi={10.1016/j.simpa.2024.100657},
   publisher={Elsevier}
 }
 
@@ -135,6 +136,7 @@
   number={4},
   pages={1507--1521},
   year={2020},
+  doi={10.1007/s12145-020-00506-0},
   publisher={Springer}
 }
 
@@ -152,6 +154,7 @@
 	date = {2019-11-13},
 	year = {2019},
 	month = {11},
+	doi={10.21105/joss.01886},
 	day = {13},
 }
 
