### 2025-09-23 151be3c
revision paper
