+++ joos_paper/paper.bib
@@ -1,7 +1,7 @@
-@article{gillies2013rasterio,
+@misc{gillies2013rasterio,
   title={Rasterio: Geospatial raster I/O for Python programmers},
   author={Gillies, Sean and Ward, B and Petersen, AS and others},
-  journal={URL https://github. com/mapbox/rasterio},
+  journal={URL https://github.com/mapbox/rasterio},
   doi={},
   year={2013}
 }
@@ -83,14 +83,14 @@
 }
 
 
-@article{gillies2013shapely,
+@misc{gillies2013shapely,
   title={The shapely user manual},
   author={Gillies, Sean and others},
-  journal={URL https://pypi. org/project/Shapely},
+  journal={https://pypi.org/project/Shapely},
   year={2013}
 }
 
-@article{jordahl2021geopandas,
+@misc{jordahl2021geopandas,
   title={geopandas/geopandas: v0. 7.0},
   author={Jordahl, Kelsey and Van den Bossche, Joris and Wasserman, Jacob and McBride, James and Fleischmann, Martin and Gerard, Jeffrey and Tratner, Jeff and Perry, Matthew and Farmer, Carson and Hjelle, Geir Arne and others},
   journal={Zenodo},
