### 2025-09-23 fd789c6
final modif in paper
