+++ README.md
@@ -4,6 +4,8 @@
 Rastereasy
 ==========
 
+Click on the logo to access the complete documentation.
+
 [![Click here for a complete documentation](./illus/logo.png)](https://rastereasy.github.io/)
 
 
