### 2025-10-17 d8c7be8
Bump version to 0.3.0
