+++ src/rastereasy/utils.py
@@ -24,15 +24,11 @@ from matplotlib.widgets import Button
 import ipywidgets as widgets
 import warnings
 if os.environ.get('DISPLAY', '') == '':
-    # Utilise 'agg' si aucune interface graphique n'est détectée
     matplotlib.use('agg')
 
 from scipy import ndimage, signal
 import json
 
-#else:
-    # Utilise 'tkagg' pour un affichage interactif standard
-#    matplotlib.use('tkagg')
 
 
 
@@ -1216,7 +1212,6 @@ def reset_matplotlib(mode='inline'):
             print(f"Warning: Impossible to run run_line_magic : {e}")
 
 
-
 def is_notebook():
     try:
         from IPython import get_ipython
@@ -1224,17 +1219,187 @@ def is_notebook():
     except:
         return False
 
+import matplotlib
+import matplotlib.pyplot as plt
+from matplotlib.widgets import Button
+import sys
 
-def plot_clic_spectra(im, imc, figsize=(15, 5), plot_legend=False, names=None,
+# You will need to install a library for the Qt backend.
+# In your terminal: pip install pyqt5
+
+def is_notebook():
+    """Checks if the code is running in a Jupyter-like environment."""
+    try:
+        shell = get_ipython().__class__.__name__
+        if shell == 'ZMQInteractiveShell':
+            return True   # Jupyter notebook or qtconsole
+        elif shell == 'TerminalInteractiveShell':
+            return False  # Terminal IPython
+        else:
+            return False  # Other type (?)
+    except NameError:
+        return False      # Probably a standard Python interpreter
+
+class SpectraCollector:
+    """
+    A class to manage an interactive Matplotlib window for collecting
+    spectra by clicking on an image.
+    """
+    def __init__(self, im, imc, figsize=(15, 5), names=None, title_im="Click on the image", 
+                 title_spectra="Spectra", xlabel="Bands", ylabel="Value", plot_legend=False,
+                 offset_i=0, offset_j=0):
+        
+        # --- Backend Configuration ---
+        # Force an interactive backend if not in a notebook.
+        # Jupyter's 'widget' backend must be enabled beforehand with %matplotlib widget.
+        if is_notebook():
+            ipython = get_ipython()
+            ipython.run_line_magic('matplotlib', 'inline')
+            ipython.run_line_magic('matplotlib', 'widget')
+        else:
+            plt.ion()  # Interactive mode for the standard shell
+
+        plt.close('all')
+
+        if not is_notebook():
+            try:
+                matplotlib.use('Qt5Agg')
+            except ImportError:
+                print("Warning: PyQt5 is not installed. Interactivity might not work.")
+                print("Please install it with: pip install pyqt5")
+
+        # --- Initialize variables ---
+        self.im = im
+        self.imc = imc
+        self.names = names
+        self.plot_legend = plot_legend
+        self.offset_i = offset_i
+        self.offset_j = offset_j
+        self.end_collect = False
+        self.series_spectrales = []
+        self.val_i = []
+        self.val_j = []
+        
+        # --- Create the figure ---
+        plt.close('all')
+        self.fig, (self.ax1, self.ax2) = plt.subplots(1, 2, figsize=figsize)
+        
+        self.ax1.imshow(self.imc, extent=[offset_j, self.imc.shape[1] + offset_j, self.imc.shape[0] + offset_i, offset_i])
+        self.ax1.set_title(title_im)
+        self.ax2.set_title(title_spectra)
+        self.ax2.set_xlabel(xlabel)
+        self.ax2.set_ylabel(ylabel)
+
+        # --- Add widgets ---
+        self.close_button = Button(plt.axes([0.04, 0.92, 0.1, 0.05]), 'Finish')
+        self.close_button.on_clicked(self.close_figure)
+
+        # --- Connect the click event ---
[…]
