+++ src/rastereasy/rastereasy.py
@@ -5,6 +5,9 @@ Contains the main Geoimage class and utility functions for raster processing.
 
 import os
 # Google Colab detection and setup
+import matplotlib
+
+
 try:
     import google.colab
     IN_COLAB = True
@@ -17,6 +20,7 @@ try:
         os.system('pip install ipympl')
 except ImportError:
     IN_COLAB = False
+    matplotlib.use('Qt5Agg') #  'Qt5Agg' 'TkAgg'
 end_collect = False  # Flag used for collecting spectra
 
 # Standard library imports
@@ -33,7 +37,6 @@ from sklearn.cluster import KMeans
 from sklearn.preprocessing import StandardScaler, MinMaxScaler
 import matplotlib.pyplot as plt
 from matplotlib.backends.backend_webagg_core import NavigationToolbar2WebAgg
-import matplotlib
 
 # Rasterio and geospatial imports
 import rasterio as rio
@@ -1245,7 +1248,7 @@ class Visualizer:
     @staticmethod
     def plot_spectra(im, bands=None, fig_size=(15, 5), percentile=2, title='',
                      title_im="Original image (click outside to stop)",
-                     title_spectra="Spectra", xlabel="Bands", ylabel="Value"):
+                     title_spectra="Spectra", xlabel="Bands", ylabel="Value", offset_i=0,offset_j=0):
         """
         Plots and extracts spectral values from user-selected pixels on a multispectral image.
 
@@ -1277,6 +1280,12 @@ class Visualizer:
         ylabel : str, optional
             Y-axis label for the spectra curves plot.
             Default is "Value".
+        offset_i : int, optional
+            Offset to add to i coordinates (in case of a zoom)
+            Default is 0.
+        offset_j : int, optional
+            Offset to add to j coordinates (in case of a zoom)
+            Default is 0.
 
         Returns
         -------
@@ -1308,7 +1317,6 @@ class Visualizer:
         def on_end_collect(series, val_i, val_j):
             # Code to be executed after the user finishes selecting pixels
             print("User has finished selecting pixels.")
-
         series, val_i, val_j, end_collect = plot_clic_spectra(
             im.numpy_channel_last(),
             imc,
@@ -1318,7 +1326,10 @@ class Visualizer:
             title_spectra=title_spectra,
             xlabel=xlabel,
             ylabel=ylabel,
-            callback=on_end_collect
+            callback=on_end_collect,
+            offset_i=offset_i,
+            offset_j=offset_j,
+            colab=IN_COLAB
         )
 
         return series, val_i, val_j
@@ -1733,14 +1744,14 @@ def files2stack(imagefile_path, resolution=None, names="origin", dest_name=None,
             message="im %s has %d bands"%(imagefile_path[0],im.nb_bands)
             warnings.warn(message,category=UserWarning)
         if im.resolution != resolution:
-            im.resampling(resolution, inplace=True)
+            im.resample(resolution, inplace=True)
 
         # Process remaining images
         for i in range(len(imagefile_path) - 1):
             # Load and resample next image
             im_tmpo = Geoimage(imagefile_path[i + 1])
             if im_tmpo.resolution != resolution:
-                im_tmpo.resampling(resolution, inplace=True)
+                im_tmpo.resample(resolution, inplace=True)
 
             if im_tmpo.nb_bands!=1:
                 message="im %s has %d bands"%(imagefile_path[i + 1],im_tmpo.nb_bands)
@@ -1842,8 +1853,8 @@ def extract_common_areas(im1, im2, resolution='min', projection=None):
             else:
                 raise ValueError(f'Error: resolution {resolution} unknown. Use "min", "max", or None.')
 
[…]
