+++ pyproject.toml
@@ -8,7 +8,7 @@ build-backend = "setuptools.build_meta"
 name = "rastereasy"
 authors = [{name = "Thomas Corpetti", email = "thomas.corpetti@gmail.com"}]
 description = "A library to ease the manipulation of geospatial images"
-version = "0.2.2"
+version = "0.3.0"
 readme = "README.md"
 requires-python = ">=3.9"
 license = "MIT"
@@ -20,7 +20,9 @@ dependencies = [
     "scikit-learn",
     "matplotlib",
     "ipywidgets",
+    "PyQt5",
     "ipympl"
 ]
 [project.urls]
-Homepage = "https://github.com/pypa/sampleproject"
+Homepage = "https://rastereasy.github.io/"
+Repository = "https://github.com/pythonraster/rastereasy"
