+++ src/rastereasy.egg-info/SOURCES.txt
@@ -1,18 +1,21 @@
+CONTRIBUTING.md
 LICENCE
 README.md
 environment_rastereasy.yml
 pyproject.toml
 .github/workflows/draft_pdf.yml
 .github/workflows/tests.yml
+dist/rastereasy-0.3.0-py3-none-any.whl
+dist/rastereasy-0.3.0.tar.gz
+illus/logo.png
 illus/spectra.png
 joos_paper/harmonization.jpg
 joos_paper/illustrations.jpg
 joos_paper/illustrations.pdf
 joos_paper/illustrations2.jpg
+joos_paper/illustrations3.jpg
 joos_paper/paper.bib
 joos_paper/paper.md
-src/.DS_Store
-src/rastereasy/.DS_Store
 src/rastereasy/__init__.py
 src/rastereasy/rastereasy.py
 src/rastereasy/utils.py
@@ -21,11 +24,6 @@ src/rastereasy.egg-info/SOURCES.txt
 src/rastereasy.egg-info/dependency_links.txt
 src/rastereasy.egg-info/requires.txt
 src/rastereasy.egg-info/top_level.txt
-src/rastereasy/.ipynb_checkpoints/Untitled-checkpoint.ipynb
-src/rastereasy/.ipynb_checkpoints/rastereasy-checkpoint.py
-src/rastereasy/.ipynb_checkpoints/utils-checkpoint.py
-src/rastereasy/__pycache__/rastereasy.cpython-311.pyc
-src/rastereasy/__pycache__/utils.cpython-311.pyc
 tests/test_core.py
 tests/data/RGB_common_1.tif
 tests/data/RGB_common_2.tif
