### 2025-11-13 3e6a922
Fix typos and improve clarity in paper.md
