### 2025-11-13 #8666 @hugoledoux
Sorry for the delay, beginning of the teaching quarter here and I was a bit too busy for my liking...

OK for the authors, it's a bit unorthodox that all commits are made by one person but it's fine for the paper.

I have proof-read the paper and made several small changes, they are in that PR: https://github.com/pythonraster/rastereasy/pull/1

I removed the "an" in the title because I don't think it's necessary, but ultimately it's up to you.

You didn't "protect" in you BibTeX words that ought to be capitalised, eg Python. I added {} to protect them, but please have another go at the bibtex and check that in the PDF they are correctly capitalised.

Oh, and weirdly, the folder where you have the JOSS paper is called `joos_paper`... not super important but I would fix this.

Once you've done that, we can start the acceptance process (and I will not take 3w to respond, I promise!).
