### 2025-11-13 fb9d780
final changes proposed by hugoledoux
