+++ paper/paper.bib
@@ -1,0 +1,159 @@
+@misc{gillies2013rasterio,
+  title={Rasterio: Geospatial raster I/O for {Python} programmers},
+  author={Gillies, Sean and Ward, B and Petersen, AS and others},
+  journal={URL https://github.com/mapbox/rasterio},
+  doi={},
+  year={2013}
+}
+
+@incollection{mamatov2024geospatial,
+  title={Geospatial Data Storage and Management},
+  author={Mamatov, Islom and Galety, Mohammad Gouse and Alimov, Ruslan and Sriharsha, AV and Rofoo, Fanar Fareed Hanna and Sunitha, Gurram},
+  booktitle={Ethics, Machine Learning, and {Python} in Geospatial Analysis},
+  pages={150--167},
+  year={2024},
+  doi={10.4018/979-8-3693-6381-2.ch007},
+  publisher={IGI Global Scientific Publishing}
+}
+
+
+
+@article{ritter1997geotiff,
+  title={The {GeoTiff} data interchange standard for raster geographic images},
+  author={Ritter, Niles and Ruth, Mike},
+  journal={International Journal of Remote Sensing},
+  volume={18},
+  number={7},
+  pages={1637--1647},
+  year={1997},
+  doi={10.1080/014311697218340},
+  publisher={Taylor \& Francis}
+}
+@article{xue2017significant,
+  title={Significant remote sensing vegetation indices: A review of developments and applications},
+  author={Xue, Jinru and Su, Baofeng},
+  journal={Journal of sensors},
+  volume={2017},
+  number={1},
+  pages={1353691},
+  year={2017},
+  doi={10.1155/2017/1353691},
+  publisher={Wiley Online Library}
+}
+@inproceedings{courty2016optimal,
+  title={Optimal transport for data fusion in remote sensing},
+  author={Courty, Nicolas and Flamary, R{\'e}mi and Tuia, Devis and Corpetti, Thomas},
+  booktitle={2016 IEEE international geoscience and remote sensing symposium (IGARSS)},
+  pages={3571--3574},
+  year={2016},
+  doi={10.1109/IGARSS.2016.7729925},
+  organization={IEEE}
+}
+
+@article{ikotun2023k,
+  title={K-means clustering algorithms: A comprehensive review, variants analysis, and advances in the era of big data},
+  author={Ikotun, Abiodun M and Ezugwu, Absalom E and Abualigah, Laith and Abuhaija, Belal and Heming, Jia},
+  journal={Information Sciences},
+  volume={622},
+  pages={178--210},
+  year={2023},
+  doi={10.1016/j.ins.2022.11.139},
+  publisher={Elsevier}
+}
+
+@article{shafer1992dempster,
+  title={Dempster-shafer theory},
+  author={Shafer, Glenn},
+  journal={Encyclopedia of artificial intelligence},
+  volume={1},
+  pages={330--331},
+  year={1992}
+}
+
+@article{harris2020array,
+  title={Array programming with NumPy},
+  author={Harris, Charles R and Millman, K Jarrod and Van Der Walt, St{\'e}fan J and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J and others},
+  journal={nature},
+  volume={585},
+  number={7825},
+  pages={357--362},
+  year={2020},
+  doi={10.1038/s41586-020-2649-2},
+  publisher={Nature Publishing Group UK London}
+}
+
+
+@misc{gillies2013shapely,
+  title={The {Shapely} user manual},
+  author={Gillies, Sean and others},
[…]
