+++ joos_paper/paper.md
@@ -1,118 +1,0 @@
----
-title: 'Rastereasy: A Python package for easy manipulation of remote sensing images'
-tags:
-- Python
-- Remote sensing
-- Geospatial analysis
-- Image processing
-- GIS
-authors:
-- name: Thomas Corpetti
-  orcid: 0000-0002-0257-138X
-  corresponding: true
-  affiliation: 1
-- name: Pierrick Matelot
-  affiliation: 2
-- name: Augustin de la Brosse
-  affiliation: 1
-- name: Candide Lissak
-  orcid: 0000-0003-3393-7767
-  affiliation: 2
-affiliations:
-- name: 'CNRS, UMR 6554 LETG, Univ. Rennes 2, Place du Recteur Henri Le Moal, 35043 Rennes Cedex, France'
-  index: 1
-- name: 'Université de Rennes, Inserm, Irset, UMR_S 1085'
-  index: 2
-date: 03 June 2025
-bibliography: paper.bib
----
-
-# Summary
-
-Working with remote sensing data often involves managing large, multi-band georeferenced rasters with varying spatial resolutions, extents, and coordinate reference systems [@mamatov2024geospatial].
-
-`rastereasy` is a Python library designed to provide a high-level, human-readable interface for common geospatial raster and vector operations (e.g., *.tif, *.jp2, *.shp) [@ritter1997geotiff; @mamatov2024geospatial]. Built on well-established libraries including `rasterio`, `numpy`, `shapely`, `geopandas`, and `scikit-learn` [@gillies2013rasterio; @harris2020array; @gillies2013shapely; @jordahl2021geopandas; @kramer2016scikit], it enables users to perform typical GIS tasks—such as resampling, cropping, reprojection, stacking, clipping rasters with shapefiles, or rasterizing vector layers—in just a few lines of code. Some basic machine learning functionalities (clustering, fusion) are also implemented.
-
-By abstracting away much of the underlying technical complexity, `rastereasy` makes geospatial processing directly accessible within Python scripts. It is particularly suited for analysts and machine learning practitioners who need to integrate geospatial data handling into their workflows without deep GIS expertise, while also helping experienced geographers prototype more quickly. Beyond core raster operations, it includes utilities for harmonizing multi-source imagery, performing clustering and domain adaptation, and preparing datasets for downstream analysis.
-
-
-With its current implementation, `rastereasy` provides a solid foundation for further development and integration into the Python geospatial ecosystem.
-
-
-# Statement of need
-
-
-Many established remote sensing libraries such as `rasterio`, `Raster Forge`, `PODPAC`, `EarthPy` or `GDAL` [@garrard2016geoprocessing; @gillies2013rasterio; @oliveira2024raster; @ueckermann2020podpac; @Wasser2019EarthPy] provide extensive and powerful functionality for reading, writing, and processing geospatial raster data. However, they can be verbose and often require a solid understanding of geospatial concepts such as projections, data structures, coordinate reference systems, geotransforms, and metadata management. While efficient, many of these libraries are specialized in specific sub-tasks (e.g., visualization, array manipulation, or graphical interfaces) and may not fully meet the needs of users whose primary expertise lies outside GIS, such as data scientists, ecologists, agronomists, or climate researchers. As a result, the learning curve can be steep and may slow down the development of operational workflows.
-
-
[…]
