+++ src/rastereasy.egg-info/SOURCES.txt
@@ -1,3 +1,4 @@
+.zenodo.json
 CONTRIBUTING.md
 LICENCE
 README.md
@@ -5,17 +6,18 @@ environment_rastereasy.yml
 pyproject.toml
 .github/workflows/draft_pdf.yml
 .github/workflows/tests.yml
-dist/rastereasy-0.3.0-py3-none-any.whl
-dist/rastereasy-0.3.0.tar.gz
+dist/rastereasy-0.3.2-py3-none-any.whl
+dist/rastereasy-0.3.2.tar.gz
 illus/logo.png
 illus/spectra.png
-joos_paper/harmonization.jpg
-joos_paper/illustrations.jpg
-joos_paper/illustrations.pdf
-joos_paper/illustrations2.jpg
-joos_paper/illustrations3.jpg
-joos_paper/paper.bib
-joos_paper/paper.md
+paper/harmonization.jpg
+paper/illustrations.jpg
+paper/illustrations.pdf
+paper/illustrations2.jpg
+paper/illustrations3.jpg
+paper/paper.bib
+paper/paper.md
+paper/paper.pdf
 src/rastereasy/__init__.py
 src/rastereasy/rastereasy.py
 src/rastereasy/utils.py
