### 2025-11-27 5a76edf
Fix Zenodo metadata
