### 2025-11-28 #8666 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
